Split up big fat file.
