+++ datacube/cfg/exceptions.py
@@ -1,0 +1,8 @@
+# This file is part of the Open Data Cube, see https://opendatacube.org for more information
+#
+# Copyright (c) 2015-2023 ODC Contributors
+# SPDX-License-Identifier: Apache-2.0
+
+
+class ConfigException(Exception):
+    pass
